FIX : modified field exam result

+++ app/apps/register_student/models.py
@@ -31,9 +31,9 @@ class Student(models.Model):
 class ExamResults(models.Model):
     admission_no = models.CharField(max_length=20)  # Specify the max_length
     exam_name = models.CharField(max_length=100)     # Specify the max_length
-    physics = models.IntegerField(null=True, blank=True)
-    chemistry = models.IntegerField(null=True, blank=True)
-    maths = models.IntegerField(null=True, blank=True)
+    physics = models.DecimalField(null=True, blank=True,decimal_places=2,max_digits=5)
+    chemistry = models.DecimalField(null=True, blank=True,decimal_places=2,max_digits=5)
+    maths = models.DecimalField(null=True, blank=True,decimal_places=2,max_digits=5)
     
     
 class LeaderBoard(models.Model):
